# Make reactstrap compile faster