--- a/types/reactstrap/index.d.ts
+++ b/types/reactstrap/index.d.ts
@@ -16,257 +16,170 @@ export interface CSSModule {
   [className: string]: string;
 }
 
-import { default as Alert_ } from './lib/Alert';
-export class Alert<T = {}> extends Alert_<T> {}
+export { default as Alert } from './lib/Alert';
 export { AlertProps } from './lib/Alert';
-import { default as Badge_ } from './lib/Badge';
-export class Badge<T = {}> extends Badge_<T> {}
+export { default as Badge } from './lib/Badge';
 export { BadgeProps } from './lib/Badge';
-import { default as Breadcrumb_ } from './lib/Breadcrumb';
-export class Breadcrumb<T = {}> extends Breadcrumb_<T> {}
+export { default as Breadcrumb } from './lib/Breadcrumb';
 export { BreadcrumbProps } from './lib/Breadcrumb';
-import { default as BreadcrumbItem_ } from './lib/BreadcrumbItem';
-export class BreadcrumbItem<T = {}> extends BreadcrumbItem_<T> {}
+export { default as BreadcrumbItem } from './lib/BreadcrumbItem';
 export { BreadcrumbItemProps } from './lib/BreadcrumbItem';
-import { default as Button_ } from './lib/Button';
-export class Button<T = {}> extends Button_<T> {}
+export { default as Button } from './lib/Button';
 export { ButtonProps } from './lib/Button';
-import { default as ButtonDropdown_ } from './lib/ButtonDropdown';
-export class ButtonDropdown<T = {}> extends ButtonDropdown_<T> {}
+export { default as ButtonDropdown } from './lib/ButtonDropdown';
 export { ButtonDropdownProps } from './lib/ButtonDropdown';
-import { default as ButtonGroup_ } from './lib/ButtonGroup';
-export class ButtonGroup<T = {}> extends ButtonGroup_<T> {}
+export { default as ButtonGroup } from './lib/ButtonGroup';
 export { ButtonGroupProps } from './lib/ButtonGroup';
-import { default as ButtonToolbar_ } from './lib/ButtonToolbar';
-export class ButtonToolbar<T = {}> extends ButtonToolbar_<T> {}
+export { default as ButtonToolbar } from './lib/ButtonToolbar';
 export { ButtonToolbarProps } from './lib/ButtonToolbar';
-import { default as Card_ } from './lib/Card';
-export class Card<T = {}> extends Card_<T> {}
+export { default as Card } from './lib/Card';
 export { CardProps } from './lib/Card';
-import { default as CardBody_ } from './lib/CardBody';
-export class CardBody<T = {}> extends CardBody_<T> {}
+export { default as CardBody } from './lib/CardBody';
 export { CardBodyProps } from './lib/CardBody';
-import { default as CardColumns_ } from './lib/CardColumns';
-export class CardColumns<T = {}> extends CardColumns_<T> {}
+export { default as CardColumns } from './lib/CardColumns';
 export { CardColumnsProps } from './lib/CardColumns';
-import { default as CardDeck_ } from './lib/CardDeck';
-export class CardDeck<T = {}> extends CardDeck_<T> {}
+export { default as CardDeck } from './lib/CardDeck';
 export { CardDeckProps } from './lib/CardDeck';
-import { default as CardFooter_ } from './lib/CardFooter';
-export class CardFooter<T = {}> extends CardFooter_<T> {}
+export { default as CardFooter } from './lib/CardFooter';
 export { CardFooterProps } from './lib/CardFooter';
-import { default as CardGroup_ } from './lib/CardGroup';
-export class CardGroup<T = {}> extends CardGroup_<T> {}
+export { default as CardGroup } from './lib/CardGroup';
 export { CardGroupProps } from './lib/CardGroup';
-import { default as CardHeader_ } from './lib/CardHeader';
-export class CardHeader<T = {}> extends CardHeader_<T> {}
+export { default as CardHeader } from './lib/CardHeader';
 export { CardHeaderProps } from './lib/CardHeader';
-import { default as CardImg_ } from './lib/CardImg';
-export class CardImg<T = {}> extends CardImg_<T> {}
+export { default as CardImg } from './lib/CardImg';
 export { CardImgProps } from './lib/CardImg';
-import { default as CardImgOverlay_ } from './lib/CardImgOverlay';
-export class CardImgOverlay<T = {}> extends CardImgOverlay_<T> {}
+export { default as CardImgOverlay } from './lib/CardImgOverlay';
 export { CardImgOverlayProps } from './lib/CardImgOverlay';
-import { default as CardLink_ } from './lib/CardLink';
-export class CardLink<T = {}> extends CardLink_<T> {}
+export { default as CardLink } from './lib/CardLink';
 export { CardLinkProps } from './lib/CardLink';
-import { default as CardSubtitle_ } from './lib/CardSubtitle';
-export class CardSubtitle<T = {}> extends CardSubtitle_<T> {}
+export { default as CardSubtitle } from './lib/CardSubtitle';
 export { CardSubtitleProps } from './lib/CardSubtitle';
-import { default as CardText_ } from './lib/CardText';
-export class CardText<T = {}> extends CardText_<T> {}
+export { default as CardText } from './lib/CardText';
 export { CardTextProps } from './lib/CardText';
-import { default as CardTitle_ } from './lib/CardTitle';
-export class CardTitle<T = {}> extends CardTitle_<T> {}
+export { default as CardTitle } from './lib/CardTitle';
 export { CardTitleProps } from './lib/CardTitle';
-import { default as Carousel_ } from './lib/Carousel';
-export class Carousel<T = {}> extends Carousel_<T> {}
+export { default as Carousel } from './lib/Carousel';
 export { CarouselProps } from './lib/Carousel';
-import { default as CarouselItem_ } from './lib/CarouselItem';
-export class CarouselItem<T = {}> extends CarouselItem_<T> {}
+export { default as CarouselItem } from './lib/CarouselItem';
 export { CarouselItemProps } from './lib/CarouselItem';
-import { default as CarouselControl_ } from './lib/CarouselControl';
-export class CarouselControl<T = {}> extends CarouselControl_<T> {}
+export { default as CarouselControl } from './lib/CarouselControl';
 export { CarouselControlProps } from './lib/CarouselControl';
-import { default as CarouselIndicators_ } from './lib/CarouselIndicators';
-export class CarouselIndicators<T = {}> extends CarouselIndicators_<T> {}
+export { default as CarouselIndicators } from './lib/CarouselIndicators';
 export { CarouselIndicatorsProps } from './lib/CarouselIndicators';
-import { default as CarouselCaption_ } from './lib/CarouselCaption';
-export class CarouselCaption<T = {}> extends CarouselCaption_<T> {}
+export { default as CarouselCaption } from './lib/CarouselCaption';
 export { CarouselCaptionProps } from './lib/CarouselCaption';
-import { default as Col_ } from './lib/Col';
-export class Col<T = {}> extends Col_<T> {}
+export { default as Col } from './lib/Col';
 export { ColProps } from './lib/Col';
-import { default as Collapse_ } from './lib/Collapse';
-export class Collapse<T = {}> extends Collapse_<T> {}
+export { default as Collapse } from './lib/Collapse';
 export { CollapseProps } from './lib/Collapse';
-import { default as Container_ } from './lib/Container';
-export class Container<T = {}> extends Container_<T> {}
+export { default as Container } from './lib/Container';
 export { ContainerProps } from './lib/Container';
-import { default as CustomInput_ } from './lib/CustomInput';
-export class CustomInput<T = {}> extends CustomInput_<T> {}
+export { default as CustomInput } from './lib/CustomInput';
 export { CustomInputProps } from './lib/CustomInput';
-import { default as Dropdown_ } from './lib/Dropdown';
-export class Dropdown<T = {}> extends Dropdown_<T> {}
+export { default as Dropdown } from './lib/Dropdown';
 export { DropdownProps } from './lib/Dropdown';
-import { default as DropdownItem_ } from './lib/DropdownItem';
-export class DropdownItem<T = {}> extends DropdownItem_<T> {}
+export { default as DropdownItem } from './lib/DropdownItem';
 export { DropdownItemProps } from './lib/DropdownItem';
-import { default as DropdownMenu_ } from './lib/DropdownMenu';
-export class DropdownMenu<T = {}> extends DropdownMenu_<T> {}
+export { default as DropdownMenu } from './lib/DropdownMenu';
 export { DropdownMenuProps } from './lib/DropdownMenu';
-import { default as DropdownToggle_ } from './lib/DropdownToggle';
-export class DropdownToggle<T = {}> extends DropdownToggle_<T> {}
+export { default as DropdownToggle } from './lib/DropdownToggle';
 export { DropdownToggleProps } from './lib/DropdownToggle';
-import { default as Fade_ } from './lib/Fade';
-export class Fade<T = {}> extends Fade_<T> {}
+export { default as Fade } from './lib/Fade';
 export { FadeProps } from './lib/Fade';
-import { default as Form_ } from './lib/Form';
-export class Form<T = {}> extends Form_<T> {}
+export { default as Form } from './lib/Form';
 export { FormProps } from './lib/Form';
-import { default as FormFeedback_ } from './lib/FormFeedback';
-export class FormFeedback<T = {}> extends FormFeedback_<T> {}
+export { default as FormFeedback } from './lib/FormFeedback';
 export { FormFeedbackProps } from './lib/FormFeedback';
-import { default as FormGroup_ } from './lib/FormGroup';
-export class FormGroup<T = {}> extends FormGroup_<T> {}
+export { default as FormGroup } from './lib/FormGroup';
 export { FormGroupProps } from './lib/FormGroup';
-import { default as FormText_ } from './lib/FormText';
-export class FormText<T = {}> extends FormText_<T> {}
+export { default as FormText } from './lib/FormText';
 export { FormTextProps } from './lib/FormText';
-import { default as Input_ } from './lib/Input';
-export class Input<T = {}> extends Input_<T> {}
+export { default as Input } from './lib/Input';
 export { InputProps } from './lib/Input';
-import { default as InputGroup_ } from './lib/InputGroup';
-export class InputGroup<T = {}> extends InputGroup_<T> {}
+export { default as InputGroup } from './lib/InputGroup';
 export { InputGroupProps } from './lib/InputGroup';
-import { default as InputGroupAddon_ } from './lib/InputGroupAddon';
-export class InputGroupAddon<T = {}> extends InputGroupAddon_<T> {}
+export { default as InputGroupAddon } from './lib/InputGroupAddon';
 export { InputGroupAddonProps } from './lib/InputGroupAddon';
-import { default as InputGroupButtonDropdown_ } from './lib/InputGroupButtonDropdown';
-export class InputGroupButtonDropdown<T = {}> extends InputGroupButtonDropdown_<T> {}
+export { default as InputGroupButtonDropdown } from './lib/InputGroupButtonDropdown';
 export { InputGroupButtonDropdownProps } from './lib/InputGroupButtonDropdown';
-import { default as InputGroupText_ } from './lib/InputGroupText';
-export class InputGroupText<T = {}> extends InputGroupText_<T> {}
+export { default as InputGroupText } from './lib/InputGroupText';
 export { InputGroupTextProps } from './lib/InputGroupText';
-import { default as Jumbotron_ } from './lib/Jumbotron';
-export class Jumbotron<T = {}> extends Jumbotron_<T> {}
+export { default as Jumbotron } from './lib/Jumbotron';
 export { JumbotronProps } from './lib/Jumbotron';
-import { default as Label_ } from './lib/Label';
-export class Label<T = {}> extends Label_<T> {}
+export { default as Label } from './lib/Label';
 export { LabelProps } from './lib/Label';
-import { default as ListGroup_ } from './lib/ListGroup';
-export class ListGroup<T = {}> extends ListGroup_<T> {}
+export { default as ListGroup } from './lib/ListGroup';
 export { ListGroupProps } from './lib/ListGroup';
-import { default as ListGroupItem_ } from './lib/ListGroupItem';
-export class ListGroupItem<T = {}> extends ListGroupItem_<T> {}
+export { default as ListGroupItem } from './lib/ListGroupItem';
 export { ListGroupItemProps } from './lib/ListGroupItem';
-import { default as ListGroupItemHeading_ } from './lib/ListGroupItemHeading';
-export class ListGroupItemHeading<T = {}> extends ListGroupItemHeading_<T> {}
+export { default as ListGroupItemHeading } from './lib/ListGroupItemHeading';
 export { ListGroupItemHeadingProps } from './lib/ListGroupItemHeading';
-import { default as ListGroupItemText_ } from './lib/ListGroupItemText';
-export class ListGroupItemText<T = {}> extends ListGroupItemText_<T> {}
+export { default as ListGroupItemText } from './lib/ListGroupItemText';
 export { ListGroupItemTextProps } from './lib/ListGroupItemText';
-import { default as Media_ } from './lib/Media';
-export class Media<T = {}> extends Media_<T> {}
+export { default as Media } from './lib/Media';
 export { MediaProps } from './lib/Media';
-import { default as Modal_ } from './lib/Modal';
-export class Modal<T = {}> extends Modal_<T> {}
+export { default as Modal } from './lib/Modal';
 export { ModalProps } from './lib/Modal';
-import { default as ModalBody_ } from './lib/ModalBody';
-export class ModalBody<T = {}> extends ModalBody_<T> {}
+export { default as ModalBody } from './lib/ModalBody';
 export { ModalBodyProps } from './lib/ModalBody';
-import { default as ModalFooter_ } from './lib/ModalFooter';
-export class ModalFooter<T = {}> extends ModalFooter_<T> {}
+export { default as ModalFooter } from './lib/ModalFooter';
 export { ModalFooterProps } from './lib/ModalFooter';
-import { default as ModalHeader_ } from './lib/ModalHeader';
-export class ModalHeader<T = {}> extends ModalHeader_<T> {}
+export { default as ModalHeader } from './lib/ModalHeader';
 export { ModalHeaderProps } from './lib/ModalHeader';
-import { default as Nav_ } from './lib/Nav';
-export class Nav<T = {}> extends Nav_<T> {}
+export { default as Nav } from './lib/Nav';
 export { NavProps } from './lib/Nav';
-import { default as Navbar_ } from './lib/Navbar';
-export class Navbar<T = {}> extends Navbar_<T> {}
+export { default as Navbar } from './lib/Navbar';
 export { NavbarProps } from './lib/Navbar';
-import { default as NavbarBrand_ } from './lib/NavbarBrand';
-export class NavbarBrand<T = {}> extends NavbarBrand_<T> {}
+export { default as NavbarBrand } from './lib/NavbarBrand';
 export { NavbarBrandProps } from './lib/NavbarBrand';
-import { default as NavbarToggler_ } from './lib/NavbarToggler';
-export class NavbarToggler<T = {}> extends NavbarToggler_<T> {}
+export { default as NavbarToggler } from './lib/NavbarToggler';
 export { NavbarTogglerProps } from './lib/NavbarToggler';
-import { default as NavItem_ } from './lib/NavItem';
-export class NavItem<T = {}> extends NavItem_<T> {}
+export { default as NavItem } from './lib/NavItem';
 export { NavItemProps } from './lib/NavItem';
-import { default as NavLink_ } from './lib/NavLink';
-export class NavLink<T = {}> extends NavLink_<T> {}
+export { default as NavLink } from './lib/NavLink';
 export { NavLinkProps } from './lib/NavLink';
-import { default as Pagination_ } from './lib/Pagination';
-export class Pagination<T = {}> extends Pagination_<T> {}
+export { default as Pagination } from './lib/Pagination';
 export { PaginationProps } from './lib/Pagination';
-import { default as PaginationItem_ } from './lib/PaginationItem';
-export class PaginationItem<T = {}> extends PaginationItem_<T> {}
+export { default as PaginationItem } from './lib/PaginationItem';
 export { PaginationItemProps } from './lib/PaginationItem';
-import { default as PaginationLink_ } from './lib/PaginationLink';
-export class PaginationLink<T = {}> extends PaginationLink_<T> {}
+export { default as PaginationLink } from './lib/PaginationLink';
 export { PaginationLinkProps } from './lib/PaginationLink';
-import { default as Popover_ } from './lib/Popover';
-export class Popover<T = {}> extends Popover_<T> {}
+export { default as Popover } from './lib/Popover';
 export { PopoverProps } from './lib/Popover';
-import { default as PopoverBody_ } from './lib/PopoverBody';
-export class PopoverBody<T = {}> extends PopoverBody_<T> {}
+export { default as PopoverBody } from './lib/PopoverBody';
 export { PopoverBodyProps } from './lib/PopoverBody';
-import { default as PopoverHeader_ } from './lib/PopoverHeader';
-export class PopoverHeader<T = {}> extends PopoverHeader_<T> {}
+export { default as PopoverHeader } from './lib/PopoverHeader';
 export { PopoverHeaderProps } from './lib/PopoverHeader';
-import { default as Progress_ } from './lib/Progress';
-export class Progress<T = {}> extends Progress_<T> {}
+export { default as Progress } from './lib/Progress';
 export { ProgressProps } from './lib/Progress';
-import { default as Row_ } from './lib/Row';
-export class Row<T = {}> extends Row_<T> {}
+export { default as Row } from './lib/Row';
 export { RowProps } from './lib/Row';
-import { default as Spinner_ } from './lib/Spinner';
-export class Spinner<T = {}> extends Spinner_<T> {}
+export { default as Spinner } from './lib/Spinner';
 export { SpinnerProps } from './lib/Spinner';
-import { default as TabContent_ } from './lib/TabContent';
-export class TabContent<T = {}> extends TabContent_<T> {}
+export { default as TabContent } from './lib/TabContent';
 export { TabContentProps } from './lib/TabContent';
-import { default as Table_ } from './lib/Table';
-export class Table<T = {}> extends Table_<T> {}
+export { default as Table } from './lib/Table';
 export { TableProps } from './lib/Table';
-import { default as TabPane_ } from './lib/TabPane';
-export class TabPane<T = {}> extends TabPane_<T> {}
+export { default as TabPane } from './lib/TabPane';
 export { TabPaneProps } from './lib/TabPane';
-import { default as Tag_ } from './lib/Tag';
-export class Tag<T = {}> extends Tag_<T> {}
+export { default as Tag } from './lib/Tag';
 export { TagProps } from './lib/Tag';
-import { default as Toast_ } from './lib/Toast';
-export class Toast<T = {}> extends Toast_<T> {}
+export { default as Toast } from './lib/Toast';
 export { ToastProps } from './lib/Toast';
-import { default as ToastBody_ } from './lib/ToastBody';
-export class ToastBody<T = {}> extends ToastBody_<T> {}
+export { default as ToastBody } from './lib/ToastBody';
 export { ToastBodyProps } from './lib/ToastBody';
-import { default as ToastHeader_ } from './lib/ToastHeader';
-export class ToastHeader<T = {}> extends ToastHeader_<T> {}
+export { default as ToastHeader } from './lib/ToastHeader';
 export { ToastHeaderProps } from './lib/ToastHeader';
-import { default as Tooltip_ } from './lib/Tooltip';
-export class Tooltip<T = {}> extends Tooltip_<T> {}
+export { default as Tooltip } from './lib/Tooltip';
 export { TooltipProps } from './lib/Tooltip';
-import {
-  UncontrolledAlert as UncontrolledAlert_,
-  UncontrolledButtonDropdown as UncontrolledButtonDropdown_,
-  UncontrolledDropdown as UncontrolledDropdown_,
-  UncontrolledTooltip as UncontrolledTooltip_,
-  UncontrolledCollapse as UncontrolledCollapse_,
-  UncontrolledCarousel as UncontrolledCarousel_,
-  UncontrolledPopover as UncontrolledPopover_
-} from './lib/Uncontrolled';
-export class UncontrolledAlert<T = {}> extends UncontrolledAlert_<T> {}
-export class UncontrolledButtonDropdown<T = {}> extends UncontrolledButtonDropdown_<T> {}
-export class UncontrolledDropdown<T = {}> extends UncontrolledDropdown_<T> {}
-export class UncontrolledTooltip<T = {}> extends UncontrolledTooltip_<T> {}
-export class UncontrolledCollapse<T = {}> extends UncontrolledCollapse_<T> {}
-export class UncontrolledCarousel<T = {}> extends UncontrolledCarousel_<T> {}
-export class UncontrolledPopover<T = {}> extends UncontrolledPopover_<T> {}
 export {
+  UncontrolledAlert,
+  UncontrolledButtonDropdown,
+  UncontrolledDropdown,
+  UncontrolledTooltip,
+  UncontrolledCollapse,
+  UncontrolledCarousel,
+  UncontrolledPopover,
   UncontrolledButtonDropdownProps,
   UncontrolledAlertProps,
   UncontrolledDropdownProps,