--- a/types/reactstrap/lib/Badge.d.ts
+++ b/types/reactstrap/lib/Badge.d.ts
@@ -1,13 +1,14 @@
 import * as React from 'react';
 import { CSSModule } from '../index';
 
-export type BadgeProps<T = {}> = React.HTMLAttributes<HTMLElement> & {
+export interface BadgeProps extends React.HTMLAttributes<HTMLElement> {
+  [key: string]: any;
   color?: string;
   pill?: boolean;
   tag?: React.ReactType;
   className?: string;
   cssModule?: CSSModule;
-} & T;
+}
 
-declare class Badge<T = {[key: string]: any}> extends React.Component<BadgeProps<T>> {}
+declare class Badge<T = {[key: string]: any}> extends React.Component<BadgeProps> {}
 export default Badge;