--- a/types/reactstrap/lib/BreadcrumbItem.d.ts
+++ b/types/reactstrap/lib/BreadcrumbItem.d.ts
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { CSSModule } from '../index';
 
-export type BreadcrumbItemProps<T = {}> = React.HTMLAttributes<HTMLElement> & {
+export interface BreadcrumbItemProps extends React.HTMLAttributes<HTMLElement> {
   tag?: React.ReactType;
   active?: boolean;
   className?: string;
@@ -9,7 +9,7 @@ export type BreadcrumbItemProps<T = {}> = React.HTMLAttributes<HTMLElement> & {
   // if a is passed as a string
   // this could be href
   [others: string]: any;
-} & T;
+}
 
-declare class BreadcrumbItem<T = {[key: string]: any}> extends React.Component<BreadcrumbItemProps<T>> {}
+declare class BreadcrumbItem<T = {[key: string]: any}> extends React.Component<BreadcrumbItemProps> {}
 export default BreadcrumbItem;