--- a/types/reactstrap/lib/ButtonDropdown.d.ts
+++ b/types/reactstrap/lib/ButtonDropdown.d.ts
@@ -1,10 +1,6 @@
 import * as React from 'react';
 import { UncontrolledDropdownProps, DropdownProps } from './Dropdown';
+export { UncontrolledDropdownProps as UncontrolledButtonDropdownProps, DropdownProps as ButtonDropdownProps };
 
-export type UncontrolledProps<T = {}> = UncontrolledDropdownProps<T>;
-export type UncontrolledButtonDropdownProps<T = {}> = UncontrolledProps<T>;
-
-export type ButtonDropdownProps<T = {}> = DropdownProps<T>;
-
-declare class ButtonDropdown<T = {[key: string]: any}> extends React.Component<ButtonDropdownProps<T>> {}
+declare class ButtonDropdown<T = {[key: string]: any}> extends React.Component<DropdownProps> {}
 export default ButtonDropdown;