--- a/types/reactstrap/lib/ButtonToolbar.d.ts
+++ b/types/reactstrap/lib/ButtonToolbar.d.ts
@@ -1,13 +1,14 @@
 import * as React from 'react';
 import { CSSModule } from '../index';
 
-export type ButtonToolbarProps<T = {}> = React.HTMLAttributes<HTMLElement> & {
+export interface ButtonToolbarProps extends React.HTMLAttributes<HTMLElement> {
+  [key: string]: any;
   tag?: React.ReactType;
   'aria-label'?: string;
   className?: string;
   cssModule?: CSSModule;
   role?: string;
-} & T;
+}
 
-declare class ButtonToolbar<T = {[key: string]: any}> extends React.Component<ButtonToolbarProps<T>> {}
+declare class ButtonToolbar<T = {[key: string]: any}> extends React.Component<ButtonToolbarProps> {}
 export default ButtonToolbar;