--- a/types/reactstrap/lib/CardBody.d.ts
+++ b/types/reactstrap/lib/CardBody.d.ts
@@ -1,11 +1,12 @@
 import * as React from 'react';
 import { CSSModule } from '../index';
 
-export type CardBodyProps<T = {}> = React.HTMLAttributes<HTMLElement> & {
+export interface CardBodyProps extends React.HTMLAttributes<HTMLElement> {
+  [key: string]: any;
   tag?: React.ReactType;
   className?: string;
   cssModule?: CSSModule;
-} & T;
+}
 
-declare class CardBody<T = {[key: string]: any}> extends React.Component<CardBodyProps<T>> {}
+declare class CardBody<T = {[key: string]: any}> extends React.Component<CardBodyProps> {}
 export default CardBody;