--- a/types/reactstrap/lib/CardColumns.d.ts
+++ b/types/reactstrap/lib/CardColumns.d.ts
@@ -1,11 +1,12 @@
 import * as React from 'react';
 import { CSSModule } from '../index';
 
-export type CardColumnsProps<T = {}> = React.HTMLAttributes<HTMLElement> & {
+export interface CardColumnsProps extends React.HTMLAttributes<HTMLElement> {
+  [key: string]: any;
   tag?: React.ReactType;
   className?: string;
   cssModule?: CSSModule;
-} & T;
+}
 
-declare class CardColumns<T = {[key: string]: any}> extends React.Component<CardColumnsProps<T>> {}
+declare class CardColumns<T = {[key: string]: any}> extends React.Component<CardColumnsProps> {}
 export default CardColumns;