--- a/types/reactstrap/lib/CardDeck.d.ts
+++ b/types/reactstrap/lib/CardDeck.d.ts
@@ -1,11 +1,12 @@
 import * as React from 'react';
 import { CSSModule } from '../index';
 
-export type CardDeckProps<T = {}> = React.HTMLAttributes<HTMLElement> & {
+export interface CardDeckProps extends React.HTMLAttributes<HTMLElement> {
+  [key: string]: any;
   tag?: React.ReactType;
   className?: string;
   cssModule?: CSSModule;
-} & T;
+}
 
-declare class CardDeck<T = {[key: string]: any}> extends React.Component<CardDeckProps<T>> {}
+declare class CardDeck<T = {[key: string]: any}> extends React.Component<CardDeckProps> {}
 export default CardDeck;