--- a/types/reactstrap/lib/CardFooter.d.ts
+++ b/types/reactstrap/lib/CardFooter.d.ts
@@ -1,11 +1,12 @@
 import * as React from 'react';
 import { CSSModule } from '../index';
 
-export type CardFooterProps<T = {}> = React.HTMLAttributes<HTMLElement> & {
+export interface CardFooterProps extends React.HTMLAttributes<HTMLElement> {
+  [key: string]: any;
   tag?: React.ReactType;
   className?: string;
   cssModule?: CSSModule;
-} & T;
+}
 
-declare class CardFooter<T = {[key: string]: any}> extends React.Component<CardFooterProps<T>> {}
+declare class CardFooter<T = {[key: string]: any}> extends React.Component<CardFooterProps> {}
 export default CardFooter;