This reverts the strictness part of #26094 without changing its
interface. It retains

1. string index signatures, so that arbitrary props can be passed in.
2. extending from React.Componenent, so that individual types can be subclassed.
3. Type parameters on the individual types.

However, string index signatures can no longer be avoided, and type parameters are ignored. This preserves interface compatibility -- existing code will not break from this change -- but disables excess prop checking. It's possible that a breaking change that returns the type package to its old strict behaviour can be made at the same time as a breaking change of reactstrap itself.

This change improves performance dramatically, and allows linting to succeed without running out of memory. This will unblock changes to `@types/react`, since reactstrap is tested as part of react's CI tests.

```
== Before (3.3.4000) ==
Files:              98
Lines:           63288
Nodes:          201816
Identifiers:     63784
Symbols:        961453
Types:          493230
Memory used:   601673K
I/O Read time:   0.02s
Parse time:      0.39s
Program time:    0.45s
Bind time:       0.22s
Check time:     10.03s
Total time:     10.70s

=== After 2 (3.3.4000) ===
Files:              98
Lines:           63255
Nodes:          200300
Identifiers:     63132
Symbols:        178691
Types:           80350
Memory used:   179229K
I/O Read time:   0.01s
Parse time:      0.41s
Program time:    0.47s
Bind time:       0.23s
Check time:      4.60s
Total time:      5.29s
```

After the first commit, the number of types is just over 337,000. On `typescript@next`, Before had 1.5 million types, and After had 107,000 types.