@sandersn 2019-04-17
I'm going to merge this to unblock react. Please comment if you want, though.