@mtsmfm 2019-11-12
I've found this PR breaks existing codes when we have our own custom component with `defaultProps`.

```tsx
import * as React from "react";
import { Button, ButtonProps } from "reactstrap";

type Props = {
  message: string;
} & ButtonProps;

class MyButton extends React.Component<Props> {
  public static defaultProps = {
    message: "hi"
  };

  public render() {
    const { message, ...props } = this.props;
    return <Button {...props}>{message}</Button>;
  }
}

const App = () => {
  return (
    <>
      /*
No overload matches this call.
  Overload 1 of 2, '(props: Readonly<Props>): MyButton', gave the following error.
    Property 'message' is missing in type '{}' but required in type 'Readonly<Props>'.
  Overload 2 of 2, '(props: Props, context?: any): MyButton', gave the following error.
    Property 'message' is missing in type '{}' but required in type 'Readonly<Props>'.
*/
      <MyButton />
    </>
  );
};
```

Here's the example:

https://codesandbox.io/s/typesreactstrap-with-defaultprops-lkru4