@sandersn 2019-11-12
@mtsmfm 

1. It is not intentional.
2. The PR should not have made any props *more* required than before.
3. This PR resulted in `@types/reacstrap@8.0.1`. Is your code broken on 8.0.1 and working on 8.0.0?

I'll defer to the `@types/reactstrap` contributors since they know more than me. I maintain Definitely Typed, but I don't know the specifics of any library. When I made this change I just avoided breaking tests while improving the compile time.