@FaithForHumans 2019-11-12
I think there might be a typo in your example you gave.

> ```tsx
> type Props = {
>   message: string;
> } & ButtonProps;

You've got message marked as required here, but you're omitting it later on. I see you've got a default prop for it, are you sure you didn't mean to mark `message` optional?

If this worked at some point, it makes me think there was a `react` types change with the way `defaultProps` worked and auto marked a prop as optional if it existed in `defaultProps`. But I'll agree with sandersn, I don't think this is due to a `@types/reactstrap` change.