@mtsmfm 2019-11-12
Here's the working example with @types/reactstrap 8.0.0

https://codesandbox.io/s/typesreactstrap-800-with-defaultprops-5h36u