@mtsmfm 2019-11-12
> 3\. This PR resulted in `@types/reacstrap@8.0.1`. Is your code broken on 8.0.1 and working on 8.0.0?
> 

Yes.

@types/reactstrap 8.0.1 https://codesandbox.io/s/typesreactstrap-with-defaultprops-lkru4
@types/reactstrap 8.0.0 https://codesandbox.io/s/typesreactstrap-800-with-defaultprops-5h36u

> If this worked at some point, it makes me think there was a react types change with the way defaultProps worked and auto marked a prop as optional if it existed in defaultProps. I don't think this is due to a @types/reactstrap change.

Hmm, in the examples, there's no difference other than @types/reactstrap's version.