@mtsmfm 2019-11-12
Previously, type parameter `T` for `ButtonProps` was `{}` by default but it now has `[key: string]: any`.

```tsx
export type ButtonProps<T = {}> = React.ButtonHTMLAttributes<HTMLButtonElement> & {
  ...
} & T
```

```tsx
export interface ButtonProps extends React.ButtonHTMLAttributes<HTMLButtonElement> {
  [key: string]: any;
  ...
}
```

It seems index signature breaks `defaultProps`.

https://codesandbox.io/s/defaultprops-jnecm