@mtsmfm 2019-11-12
@FaithForHumans

> are you sure you didn't mean to mark `message` optional?

Yes, it should be marked as optional by `defaultProps`.

https://codesandbox.io/s/defaultprops-s5jg2

https://www.typescriptlang.org/docs/handbook/release-notes/typescript-3-0.html#support-for-defaultprops-in-jsx