@sandersn 2019-11-12
Huh. @weswigham is that a known limitation of our defaultProps support, or a bug?