@weswigham 2019-11-12
We don't handle defaultProps, per sey, the react.d.ts does. I imagine some kind of handling could be added into it, if need be.